Major revision of serialization implementation.
There were too many compilation / run-time issues with BOOST_CLASS_EXPORT_IMPLEMENT in the cpp file, and the boost::serialization documentation is not very clear on this.

Changes:
* All serialize(...) functions are now in the header file
* All classes are now registered (with BOOST_CLASS_EXPORT) in serialization.hpp files.

It would be nicer to register the classes in their own hpp/cpp files. But at least now it works, and the code is also much cleaner.